+++ spinetoolbox/test/test_toolboxUI.py
@@ -25,9 +25,8 @@ import sys
 from PySide2.QtWidgets import QApplication, QWidget
 from PySide2.QtCore import SIGNAL, Qt, QPoint, QItemSelectionModel
 from PySide2.QtTest import QTest
-from ..ui_main import ToolboxUI
 from ..project import SpineToolboxProject
-from .mock_helpers import MockQWidget, qsettings_value_side_effect
+from .mock_helpers import create_toolboxui
 from ..config import APPLICATION_PATH
 from ..graphics_items import ProjectItemIcon, Link
 from ..project_item import RootProjectItem
@@ -52,14 +51,7 @@ class TestToolboxUI(unittest.TestCase):
     def setUp(self):
         """Overridden method. Runs before each test. Makes an instance of ToolboxUI class
         without opening previous project."""
-        with mock.patch("spinetoolbox.ui_main.JuliaREPLWidget") as mock_julia_repl, mock.patch(
-            "spinetoolbox.ui_main.PythonReplWidget"
-        ) as mock_python_repl, mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
-            # Replace Julia and Python REPLs with a QWidget so that the DeprecationWarning from qtconsole is not printed
-            mock_julia_repl.return_value = QWidget()
-            mock_python_repl.return_value = MockQWidget()  # Hack, because QWidget does not have test_push_vars()
-            mock_qsettings_value.side_effect = qsettings_value_side_effect  # override 'open previous project' setting
-            self.toolbox = ToolboxUI()
+        self.toolbox = create_toolboxui()
 
     def tearDown(self):
         """Overridden method. Runs after each test.
@@ -561,59 +553,82 @@ class TestToolboxUI(unittest.TestCase):
         self.fail()
 
     def test_tasks_before_exit_without_open_project(self):
+        """_tasks_before_exit is called with every possible combination of the two QSettings values that it uses.
+        This test is done without a project so MUT only calls QSettings.value() once.
+        This can probably be simplified but at least it does not edit user's Settings, while doing the test."""
         self.assertIsNone(self.toolbox.project())
         settings = self.toolbox.qsettings()
-        settings.setValue("appSettings/showExitPrompt", "0")
-        settings.setValue("appSettings/saveAtExit", "0")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_1
+            tasks = self.toolbox._tasks_before_exit()
+            mock_qsettings_value.assert_called_once()
+            mock_qsettings_value.assert_called_with("appSettings/showExitPrompt", defaultValue="2")
         self.assertEqual(tasks, [])
-        settings.setValue("appSettings/showExitPrompt", "2")
-        settings.setValue("appSettings/saveAtExit", "0")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_2
+            tasks = self.toolbox._tasks_before_exit()
+            mock_qsettings_value.assert_called_once()
+            mock_qsettings_value.assert_called_with("appSettings/showExitPrompt", defaultValue="2")
         self.assertEqual(tasks, ["prompt exit"])
-        settings.setValue("appSettings/showExitPrompt", "0")
-        settings.setValue("appSettings/saveAtExit", "1")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_3
+            tasks = self.toolbox._tasks_before_exit()
+            mock_qsettings_value.assert_called_once()
+            mock_qsettings_value.assert_called_with("appSettings/showExitPrompt", defaultValue="2")
         self.assertEqual(tasks, [])
-        settings.setValue("appSettings/showExitPrompt", "2")
-        settings.setValue("appSettings/saveAtExit", "1")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_4
+            tasks = self.toolbox._tasks_before_exit()
+            mock_qsettings_value.assert_called_once()
+            mock_qsettings_value.assert_called_with("appSettings/showExitPrompt", defaultValue="2")
         self.assertEqual(tasks, ["prompt exit"])
-        settings.setValue("appSettings/showExitPrompt", "0")
-        settings.setValue("appSettings/saveAtExit", "2")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_5
+            tasks = self.toolbox._tasks_before_exit()
+            mock_qsettings_value.assert_called_once()
+            mock_qsettings_value.assert_called_with("appSettings/showExitPrompt", defaultValue="2")
         self.assertEqual(tasks, [])
-        settings.setValue("appSettings/showExitPrompt", "2")
-        settings.setValue("appSettings/saveAtExit", "2")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_6
+            tasks = self.toolbox._tasks_before_exit()
+            mock_qsettings_value.assert_called_once()
+            mock_qsettings_value.assert_called_with("appSettings/showExitPrompt", defaultValue="2")
         self.assertEqual(tasks, ["prompt exit"])
 
     def test_tasks_before_exit_with_open_project(self):
+        """_tasks_before_exit is called with every possible combination of the two QSettings values that it uses.
+        This test is done with a 'mock' project so MUT calls QSettings.value() twice."""
         self.toolbox._project = 1  # Just make sure project is not None
         settings = self.toolbox.qsettings()
-        settings.setValue("appSettings/showExitPrompt", "0")
-        settings.setValue("appSettings/saveAtExit", "0")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_1
+            tasks = self.toolbox._tasks_before_exit()
+            self.assertEqual(2, mock_qsettings_value.call_count)
         self.assertEqual(tasks, [])
-        settings.setValue("appSettings/showExitPrompt", "2")
-        settings.setValue("appSettings/saveAtExit", "0")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_2
+            tasks = self.toolbox._tasks_before_exit()
+            self.assertEqual(2, mock_qsettings_value.call_count)
         self.assertEqual(tasks, ["prompt exit"])
-        settings.setValue("appSettings/showExitPrompt", "0")
-        settings.setValue("appSettings/saveAtExit", "1")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_3
+            tasks = self.toolbox._tasks_before_exit()
+            self.assertEqual(2, mock_qsettings_value.call_count)
         self.assertEqual(tasks, ["prompt save"])
-        settings.setValue("appSettings/showExitPrompt", "2")
-        settings.setValue("appSettings/saveAtExit", "1")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_4
+            tasks = self.toolbox._tasks_before_exit()
+            self.assertEqual(2, mock_qsettings_value.call_count)
         self.assertEqual(tasks, ["prompt save"])
-        settings.setValue("appSettings/showExitPrompt", "0")
-        settings.setValue("appSettings/saveAtExit", "2")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_5
+            tasks = self.toolbox._tasks_before_exit()
+            self.assertEqual(2, mock_qsettings_value.call_count)
         self.assertEqual(tasks, ["save"])
-        settings.setValue("appSettings/showExitPrompt", "2")
-        settings.setValue("appSettings/saveAtExit", "2")
-        tasks = self.toolbox._tasks_before_exit()
+        with mock.patch("spinetoolbox.ui_main.QSettings.value") as mock_qsettings_value:
+            mock_qsettings_value.side_effect = self._tasks_before_exit_scenario_6
+            tasks = self.toolbox._tasks_before_exit()
+            self.assertEqual(2, mock_qsettings_value.call_count)
         self.assertEqual(tasks, ["prompt exit", "save"])
 
     def test_propose_item_name(self):
@@ -732,6 +747,48 @@ class TestToolboxUI(unittest.TestCase):
         # Map scene coordinates to graphics view viewport coordinates
         qpoint = gv.mapFromScene(qpointf)  # Returns a point in Graphics view viewport coordinate system
         return qpoint
+
+    @staticmethod
+    def _tasks_before_exit_scenario_1(key, defaultValue="2"):
+        if key == "appSettings/showExitPrompt":
+            return "0"
+        elif key == "appSettings/saveAtExit":
+            return "0"
+
+    @staticmethod
+    def _tasks_before_exit_scenario_2(key, defaultValue="2"):
+        if key == "appSettings/showExitPrompt":
+            return "2"
+        elif key == "appSettings/saveAtExit":
+            return "0"
+
+    @staticmethod
+    def _tasks_before_exit_scenario_3(key, defaultValue="2"):
+        if key == "appSettings/showExitPrompt":
+            return "0"
+        elif key == "appSettings/saveAtExit":
+            return "1"
+
+    @staticmethod
+    def _tasks_before_exit_scenario_4(key, defaultValue="2"):
+        if key == "appSettings/showExitPrompt":
+            return "2"
+        elif key == "appSettings/saveAtExit":
+            return "1"
+
+    @staticmethod
+    def _tasks_before_exit_scenario_5(key, defaultValue="2"):
+        if key == "appSettings/showExitPrompt":
+            return "0"
+        elif key == "appSettings/saveAtExit":
+            return "2"
+
+    @staticmethod
+    def _tasks_before_exit_scenario_6(key, defaultValue="2"):
+        if key == "appSettings/showExitPrompt":
+            return "2"
+        elif key == "appSettings/saveAtExit":
+            return "2"
 
 
 if __name__ == '__main__':